+++ apps/simpleFoam/view/config/content/10_Discretization.Options/Gradients.qml
@@ -9,7 +9,7 @@ Column {
 
     TabsCard3 {
         model: {
-            switch (app.turbulenceModel) {
+            switch (app.turbulence.model) {
             case 'laminar':
                 return [
                         'Default',